+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
-* Added a `RenderConfig`` for configuring the Render object.
+
+* Added and formatted `Shader` and `Camera`.
+* Added a `RenderConfig` for configuring the Render object.
 * Added the `shaders` and the `Shader` class  in the shader folder.
 * Created the `Camera`, `metrics` and `gl` functions and classes (most copied from the view2).
 * Adjustment of the `ControllerConfig` and `ViewerConfig``.